Added Auth

+++ routes/web.php
@@ -8,28 +8,40 @@ use App\Http\Controllers\ListingController;
 Route::get('/', [ListingController::class, 'index']);
 
 //Show Create Form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 //Store Listing Data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 //Show Edit Form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
-//Store Edited Form
-Route::put('/listings/{listing}/', [ListingController::class, 'update']);
+//Update Listing
+Route::put('/listings/{listing}/', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete Listing
-Route::delete('/listings/{listing}/', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}/', [ListingController::class, 'destroy'])->middleware('auth');
+
+//Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 //Listing by Id
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 //Show Register Create Form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 //Store Created User
 Route::post('/users', [UserController::class, 'store']);
 
+//Show Login Form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+//Log In User
+Route::post('/users/auth', [UserController::class, 'auth']);
+
+//Log User Out
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 
+
